requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2

# Additional dependencies for production
gunicorn==21.2.0
python-dotenv==1.0.0

# Database - Supabase (PostgreSQL)
supabase==2.3.4
psycopg2-binary==2.9.7

# Authentication (for future use)
# Flask-Login==0.6.3
# Flask-WTF==1.1.1
# WTForms==3.0.1

# Email (for future use)
# Flask-Mail==0.9.1
